pyproject.toml
[project]
name = "stomlins-dash"
version = "2.0.0"
description = "stomlins.com — Astro static site. Python env for asset generation only."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "seasonal-spirals[matplotlib]>=0.3.1",
]

[tool.uv]
package = false
